Former-commit-id: 67d58c3192fec261791b8041a4d186b32c9e0af7

+++ backend/src/main/java/com/ssafy/backend/districtReport/model/vo/ClosingRateCard.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class ClosingRateCard {
 
-    int closures;
+    double closures;
     int numOfCafe;
     int numOfCafe1;
     int numOfCafe2;
